LPD-68467 Move FilePreview component to common folder
