+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/modal/FilePreviewerModalContent.tsx
@@ -8,8 +8,8 @@ import ClayLink from '@clayui/link';
 import ClayModal from '@clayui/modal';
 import React from 'react';
 
+import FilePreview from '../../common/components/FilePreview';
 import {IAssetFile} from '../../common/types/AssetType';
-import FilePreview from './asset_navigation_view/FilePreview';
 
 interface FilePreviewerModalContentProps {
 	file: IAssetFile;
